web_ui/requirements.txt
gradio
torch
torchvision
google-cloud-aiplatform
grpcio-tools